refactor(webauthn): split WebAuthnClient into partials by concern
WebAuthnClient.cs was over 1200 lines and mixed both ceremonies, PIN/UV token
acquisition, validation, and error mapping in one file. Split it into partial
files along those seams. WebAuthnClient.cs keeps construction, disposal, and
the shared zeroing helpers.

Pure move: the multiset of non-blank class-body lines is identical before and
after, and the WebAuthn unit suite reports 199 passing on both sides.
