+++ src/WebAuthn/src/Client/WebAuthnClientOptions.cs
